# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Decimal": {
				"type": "string"
			},
			"_36_Enums.TriagePriority": {
				"type": "string",
				"enum": [
					"RED",
					"YELLOW",
					"GREEN",
					"BLACK"
				]
			},
			"DefaultSelection_Prisma._36_VitalSignsPayload_": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"recordedById": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/_36_Enums.TriagePriority"
					},
					"bmi": {
						"$ref": "#/components/schemas/Decimal"
					},
					"height": {
						"$ref": "#/components/schemas/Decimal"
					},
					"weight": {
						"$ref": "#/components/schemas/Decimal"
					},
					"spo2": {
						"type": "number",
						"format": "double"
					},
					"respiratoryRate": {
						"type": "number",
						"format": "double"
					},
					"pulseRate": {
						"type": "number",
						"format": "double"
					},
					"diastolicBP": {
						"type": "number",
						"format": "double"
					},
					"systolicBP": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"$ref": "#/components/schemas/Decimal"
					},
					"visitId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"recordedById",
					"notes",
					"priority",
					"bmi",
					"height",
					"weight",
					"spo2",
					"respiratoryRate",
					"pulseRate",
					"diastolicBP",
					"systolicBP",
					"temperature",
					"visitId",
					"id"
				],
				"type": "object"
			},
			"VitalSigns": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_VitalSignsPayload_",
				"description": "Model VitalSigns"
			},
			"TriagePriority": {
				"$ref": "#/components/schemas/_36_Enums.TriagePriority"
			},
			"CreateVitalSignsInput": {
				"properties": {
					"visitId": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"systolicBP": {
						"type": "number",
						"format": "double"
					},
					"diastolicBP": {
						"type": "number",
						"format": "double"
					},
					"pulseRate": {
						"type": "number",
						"format": "double"
					},
					"respiratoryRate": {
						"type": "number",
						"format": "double"
					},
					"spo2": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"priority": {
						"$ref": "#/components/schemas/TriagePriority"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"visitId",
					"temperature",
					"systolicBP",
					"diastolicBP",
					"pulseRate",
					"respiratoryRate",
					"spo2",
					"weight",
					"height"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Omit_CreateVitalSignsInput.visitId__": {
				"properties": {
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"systolicBP": {
						"type": "number",
						"format": "double"
					},
					"diastolicBP": {
						"type": "number",
						"format": "double"
					},
					"pulseRate": {
						"type": "number",
						"format": "double"
					},
					"respiratoryRate": {
						"type": "number",
						"format": "double"
					},
					"spo2": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"priority": {
						"$ref": "#/components/schemas/_36_Enums.TriagePriority"
					},
					"notes": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateVitalSignsInput": {
				"$ref": "#/components/schemas/Partial_Omit_CreateVitalSignsInput.visitId__"
			},
...
			"_36_Enums.Gender": {
				"type": "string",
				"enum": [
					"MALE",
					"FEMALE",
					"OTHER"
				]
			},
			"Gender": {
				"$ref": "#/components/schemas/_36_Enums.Gender"
			},
			"PatientDTO": {
				"description": "Standard Patient entity projection matching Prisma return types exactly.\nPrisma returns `null` for optional database fields, not `undefined`.",
				"properties": {
					"id": {
						"type": "string"
					},
					"mrn": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"middleName": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"emergencyContactName": {
						"type": "string",
						"nullable": true
					},
					"emergencyContactPhone": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"mrn",
					"firstName",
					"lastName",
					"middleName",
					"gender",
					"dateOfBirth",
					"phone",
					"emergencyContactName",
					"emergencyContactPhone",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PatientDTO_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PatientDTO"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatientResponseDTO": {
				"$ref": "#/components/schemas/ApiResponse_PatientDTO_"
			},
			"RegisterPatientDTO": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"middleName": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"dateOfBirth": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"emergencyContactName": {
						"type": "string"
					},
					"emergencyContactPhone": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"gender",
					"dateOfBirth"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DefaultSelection_Prisma._36_PatientPayload_": {
				"properties": {
					"emergencyContactPhone": {
						"type": "string"
					},
					"emergencyContactName": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"gender": {
						"$ref": "#/components/schemas/_36_Enums.Gender"
					},
					"mrn": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"middleName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"emergencyContactPhone",
					"emergencyContactName",
					"dateOfBirth",
					"gender",
					"mrn",
					"phone",
					"middleName",
					"lastName",
					"firstName",
					"updatedAt",
					"createdAt",
					"id"
				],
				"type": "object"
			},
			"Patient": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_PatientPayload_",
				"description": "Model Patient"
			},
			"PatientListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Patient"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"count",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPrevPage": {
						"type": "boolean"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"totalPages",
					"hasNextPage",
					"hasPrevPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_PatientDTO_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PatientDTO"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"success",
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedPatientsResponseDTO": {
				"$ref": "#/components/schemas/PaginatedResponse_PatientDTO_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"description": "Authentication and RBAC Microservice",
		"contact": {}
	},
	"paths": {
		"/api/v1/vital-signs": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/VitalSigns"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Patient Visit Not Found"
					},
					"409": {
						"description": "Vital Signs Already Recorded"
					}
				},
				"tags": [
					"Vital Signs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVitalSignsInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"meta": {
											"properties": {
												"totalPages": {
													"type": "number",
													"format": "double"
												},
												"limit": {
													"type": "number",
													"format": "double"
												},
												"page": {
													"type": "number",
													"format": "double"
												},
												"total": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"totalPages",
												"limit",
												"page",
												"total"
											],
											"type": "object"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/VitalSigns"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"meta",
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Vital Signs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "patientId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "priority",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TriagePriority"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/vital-signs/visit/{visitId}": {
			"get": {
				"operationId": "GetByVisitId",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/VitalSigns"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Vital Signs Not Found"
					}
				},
				"tags": [
					"Vital Signs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "visitId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/vital-signs/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/VitalSigns"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Vital Signs Record Not Found"
					}
				},
				"tags": [
					"Vital Signs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateVitalSignsInput"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/visits/patient/{mrn}": {
			"get": {
				"operationId": "GetVisitsByPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"count": {},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"count",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Patient Visits"
				],
				"security": [
					{
						"jwt": [
							"NURSE",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mrn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/patients/register": {
			"post": {
				"operationId": "RegisterPatient",
				"responses": {
					"201": {
						"description": "Patient Registered Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatientResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request / Validation Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Patient Management"
				],
				"security": [
					{
						"jwt": [
							"NURSE",
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterPatientDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/patients/lookup": {
			"get": {
				"operationId": "LookupPatients",
				"responses": {
					"200": {
						"description": "Patients Fetched Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatientListResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - Missing query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatientListResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve patient details by Medical Record Number (MRN).\nChecks Redis cache first before falling back to PostgreSQL.",
				"tags": [
					"Patient Management"
				],
				"security": [
					{
						"jwt": [
							"NURSE",
							"DOCTOR",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "mrn",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "firstName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "lastName",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/patients": {
			"get": {
				"operationId": "GetPatients",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedPatientsResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Query Parameters"
					}
				},
				"description": "Search and list patients with pagination and filtering.",
				"tags": [
					"Patient Management"
				],
				"security": [
					{
						"jwt": [
							"NURSE",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "gender",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"MALE",
								"FEMALE",
								"OTHER"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"createdAt",
								"lastName",
								"mrn"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
...
		"/api/v1/billing/insurance/patient/{patientId}": {
			"get": {
				"operationId": "GetPatientInsurances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve active insurance policies for a patient (Cached in Redis)",
				"tags": [
					"Insurance & Claims"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "patientId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}